Reorganize/Rename the visualize unit test
